requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
pydantic
pydantic-settings
email-validator
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
